apps/rag/pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "langchain>=0.3.27",
    "langchain-core>=0.3.74",
    "langchain-huggingface>=0.3.1",
    "langchain-openai>=0.3.31",
    "langchain-postgres>=0.0.15",
    "psycopg[binary]>=3.2.9",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "sentence-transformers>=5.1.0",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]
